Move EmptyPoint to the mock folder
