+++ src/mock/cards.js
@@ -119,4 +119,17 @@ const generateCards = (amount) => {
     );
 };
 
+export const EmptyPoint = {
+  type: `taxi`,
+  city: ``,
+  startDate: null,
+  endDate: null,
+  offers: [],
+  photos: [],
+  description: ``,
+  price: 0,
+  isFavorite: false,
+  id: String(Date.now() + Math.random())
+};
+
 export const cards = generateCards(CARDS_AMOUNT);
